pyproject.toml
[project]
name = "ometer"
version = "0.1.0"
description = "OllamaMeter — list, benchmark, and compare Ollama models"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "EndoTheDev"},
]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "rich>=15.0.0",
    "python-dotenv>=1.0.0",
    "inquirerpy>=0.3.4",
]

[project.scripts]
ometer = "ometer.cli:main_entrypoint"
ollamameter = "ometer.cli:main_entrypoint"

[project.urls]
Homepage = "https://github.com/EndoTheDev/OllamaMeter"
Repository = "https://github.com/EndoTheDev/OllamaMeter"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ometer"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.2",
    "pytest-cov>=7.1.0",
]

[tool.coverage.run]
source = ["ometer"]

[tool.coverage.report]
omit = [
    "tests/*",
    "*/__main__.py",
]
